use shared session
To avoid the overhead of dialing MongoDB for every test,
we use a shared session that is returned each time.

The new NewExclusive function can be used to obtain a
separate session for tests that want to muck around with
per-session state.

We also fix a previous bug - we weren't closing the session
on Database.Close.